requirements.txt
asgiref==3.2.3
dj-database-url==0.5.0
Django==3.0.2
django-heroku==0.3.1
django-pagedown==2.0.3
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
gunicorn==20.0.4
mistune==0.8.4
psycopg2==2.8.4
PyJWT==1.7.1
pytz==2019.3
sqlparse==0.3.0
whitenoise==4.1.4
